Relax some arithmetic trait bounds on type parameter `N`
The first `impl` block contained several methods which really have
different needs as far as trait bounds on `N` are concerned, forcing the
block to have the union of all of those requirements.  By breaking this
`impl` block up we can better limit the bounds for each method to what
is strictly necessary.  This has downstream benefits in relaxing the
bounds on other trait implementations further down in the crate.

For example, `Counter::update()` does not require `N: PartialOrd` or `N:
SubAssign`, `Counter::subtract()` does not require `N: AddAssign`, and
`Counter::new()` requires neither the `PartialOrd` nor `AddAssign` nor
`SubAssign` nor `One` bounds on `N`.

This also eliminates the counter-intuitive bound `N: SubAssign` on the
implementation of `AddAssign` for `Counter<T, N>`, as well as the bound
`N: AddAssign` on the implementation of `SubAssign` for `Counter<T, N>`,
etc.