Relax some bounds unnecessarily requiring `Clone`
The implementations of `Deref` and `DerefMut` do not clone `N` and hence
do not need the bound `N: Clone`.  It looks like the bound was added in
commit cb24e7f4 because it was required by the bound `N: Clone` on the
struct itself.  The bound on the struct was later removed in 8180199a,
but it looks like the bound on these `impl`s remained.

The current implementation of `Sub<I: IntoIterator<Item = T>>` does not
clone either `T` or `N` and hence does not need either bound `T: Clone`
or `N: Clone`.  It looks like the original implementation did clone
`self`, but that was changed in cdaad73fs at which point the bound `T:
Clone` and `N: Clone` could have been removed.