Avoid unnecessary use of `Clone`
This builds on commit c2d71986, which refactored the iteration to take
ownership of the keys and values of `rhs.map` instead of references.
Not only is `value` owned and hence does not need to be cloned, but also
`key` is owned and thus there is no need to clone it to look up the
entry.