Refactor impls of `BitAnd`, `BitOr` to avoid cloning
Since we are taking ownership of both the `self` and `rhs` Counters, we
really do not have to clone either keys or values to produce new
Counters with the correct semantics for these traits.

This commit refactors these implementations to void cloning and
removes the bounds `T: Clone` and `N: Clone`.