Relax trait bound on closure in `most_common_tiebreaker()`
`slice::sort_by()` can take a closure that is `FnMut`; it does not
require `Fn`.  Thus, `Counter::most_common_tiebreaker()` could also
accept closures that are `FnMut`.

As an example, the test added by this commit will not compile with the
previous version of the code with the stricter `Fn` bound.