Simplify the implementation of `most_common_tiebreaker()`
Use match ergonomics to simplify the pattern matching.  When matching a
reference value with a non-reference pattern, the default binding mode
is automatically set to `ref`.  In this case, by removing the `&` on the
tuples and the `ref` keywords on their fields, the variables
`{a,b}_{item,count}` are bound as references.

Replace the `match` expression with `Ordering::then_with()`, which does
the same thing more succinctly.